+++ ElvUIImport.lua
@@ -9,6 +9,16 @@ function NUI:SetupLayout(layout)
 	--]]
 
 	--LAYOUT GOES HERE
+	if E.Retail then
+		E.db["WT"]["announcement"]["interrupt"]["enable"] = false
+		E.db["WT"]["combat"]["raidMarkers"]["enable"] = false
+		E.db["WT"]["item"]["extraItemsBar"]["bar2"]["anchor"] = "TOPRIGHT"
+		E.db["WT"]["item"]["extraItemsBar"]["enable"] = false
+		E.db["WT"]["misc"]["gameBar"]["enable"] = false
+		E.db["WT"]["quest"]["paragonReputation"]["text"] = "PARAGONPLUS"
+		E.db["WT"]["quest"]["switchButtons"]["font"]["name"] = "Gotham Narrow Ultra"
+		E.db["WT"]["social"]["chatBar"]["enable"] = false
+	end
 	E.db["actionbar"]["bar1"]["backdropSpacing"] = 1
 	E.db["actionbar"]["bar1"]["buttonSize"] = 28
 	E.db["actionbar"]["bar1"]["buttonSpacing"] = 1
@@ -182,7 +192,7 @@ function NUI:SetupLayout(layout)
 	E.db["auras"]["buffs"]["countFontOutline"] = "OUTLINE"
 	E.db["auras"]["buffs"]["countFontSize"] = 14
 	E.db["auras"]["buffs"]["countXOffset"] = -8
-	E.db["auras"]["buffs"]["countYOffset"] = 24
+	E.db["auras"]["buffs"]["countYOffset"] = -15
 	E.db["auras"]["buffs"]["horizontalSpacing"] = 3
 	E.db["auras"]["buffs"]["timeFont"] = "Univers"
 	E.db["auras"]["buffs"]["timeFontOutline"] = "OUTLINE"
@@ -290,7 +300,7 @@ function NUI:SetupLayout(layout)
 	E.db["general"]["backdropfadecolor"]["g"] = 0.058823529411765
 	E.db["general"]["backdropfadecolor"]["r"] = 0.058823529411765
 	E.db["general"]["bottomPanel"] = false
-	E.db["general"]["font"] = "Nimrod MT"
+	E.db["general"]["font"] = "Gotham Narrow Ultra"
 	E.db["general"]["itemLevel"]["itemLevelFont"] = "Nimrod MT"
 	E.db["general"]["loginmessage"] = false
 	E.db["general"]["lootRoll"]["nameFont"] = "Univers"
@@ -307,6 +317,18 @@ function NUI:SetupLayout(layout)
 	E.db["general"]["valuecolor"]["b"] = 0.72941017150879
 	E.db["general"]["valuecolor"]["g"] = 0.54901838302612
 	E.db["general"]["valuecolor"]["r"] = 0.95686066150665
+	E.db["mMediaTag"]["mPluginVersion"] = "2.78.1"
+	E.db["mMediaTag"]["mRoleSymbols"]["dd"] = "firenew1"
+	E.db["mMediaTag"]["mRoleSymbols"]["heal"] = "heal1"
+	E.db["mMediaTag"]["mRoleSymbols"]["tank"] = "tank1"
+	E.db["mMediaTag"]["mTags"]["afkname"] = "AFK5"
+	E.db["mMediaTag"]["mTags"]["afkpath"] = "Interface\\AddOns\\ElvUI_mMediaTag\\media\\misc\\afk5.tga"
+	E.db["mMediaTag"]["mTags"]["dndname"] = "DND5"
+	E.db["mMediaTag"]["mTags"]["dndpath"] = "Interface\\AddOns\\ElvUI_mMediaTag\\media\\misc\\dnd5.tga"
+	E.db["mMediaTag"]["mTags"]["ghostname"] = "SKULL1"
+	E.db["mMediaTag"]["mTags"]["ghostpath"] = "Interface\\AddOns\\ElvUI_mMediaTag\\media\\misc\\skull1.tga"
+	E.db["mMediaTag"]["mTags"]["skullname"] = "SKULL1"
+	E.db["mMediaTag"]["mTags"]["skullpath"] = "Interface\\AddOns\\ElvUI_mMediaTag\\media\\misc\\skull1.tga"
 	if not E.db.movers then E.db.movers = {} end
 	E.db["movers"]["BossButton"] = "BOTTOM,UIParent,BOTTOM,316,221"
 	E.db["movers"]["DurabilityFrameMover"] = "TOPRIGHT,UIParent,TOPRIGHT,-254,-202"
@@ -315,7 +337,7 @@ function NUI:SetupLayout(layout)
 	E.db["movers"]["ElvUF_RaidMover"] = "BOTTOMLEFT,UIParent,BOTTOMLEFT,4,248"
 	E.db["movers"]["ElvUF_TargetCastbarMover"] = "BOTTOM,ElvUIParent,BOTTOM,308,322"
 	E.db["movers"]["ElvUF_TargetMover"] = "BOTTOM,ElvUIParent,BOTTOM,308,340"
-	E.db["movers"]["ElvUF_TargetTargetMover"] = "BOTTOM,ElvUIParent,BOTTOM,310,275"
+	E.db["movers"]["ElvUF_TargetTargetMover"] = "BOTTOM,ElvUIParent,BOTTOM,310,285"
 	E.db["movers"]["HonorBarMover"] = "BOTTOM,ElvUIParent,BOTTOM,0,5"
 	E.db["movers"]["MicrobarMover"] = "TOP,UIParent,TOP,235,-4"
 	E.db["movers"]["ObjectiveFrameMover"] = "TOPRIGHT,UIParent,TOPRIGHT,-55,-276"
@@ -332,8 +354,13 @@ function NUI:SetupLayout(layout)
 	E.db["unitframe"]["font"] = "Gotham Narrow Ultra"
 	E.db["unitframe"]["fontOutline"] = "THICKOUTLINE"
 	E.db["unitframe"]["fontSize"] = 15
-	E.db["unitframe"]["statusbar"] = "Melli"
+	E.db["unitframe"]["statusbar"] = "mMediaTag O1"
 	E.db["unitframe"]["units"]["party"]["debuffs"]["anchorPoint"] = "LEFT"
+	E.db["unitframe"]["units"]["party"]["health"]["text_format"] = ""
+	E.db["unitframe"]["units"]["party"]["power"]["text_format"] = ""
+	E.db["unitframe"]["units"]["party"]["roleIcon"]["position"] = "LEFT"
+	E.db["unitframe"]["units"]["party"]["roleIcon"]["size"] = 25
+	E.db["unitframe"]["units"]["party"]["roleIcon"]["xOffset"] = -30
 	E.db["unitframe"]["units"]["party"]["showPlayer"] = false
 	E.db["unitframe"]["units"]["player"]["RestIcon"]["enable"] = false
 	E.db["unitframe"]["units"]["player"]["aurabar"]["enable"] = false
@@ -351,6 +378,9 @@ function NUI:SetupLayout(layout)
 	E.db["unitframe"]["units"]["player"]["power"]["detachedWidth"] = 74
 	E.db["unitframe"]["units"]["player"]["power"]["enable"] = false
 	E.db["unitframe"]["units"]["player"]["width"] = 220
+	E.db["unitframe"]["units"]["raid"]["health"]["position"] = "TOP"
+	E.db["unitframe"]["units"]["raid"]["name"]["position"] = "BOTTOM"
+	E.db["unitframe"]["units"]["raid"]["roleIcon"]["position"] = "TOPLEFT"
 	E.db["unitframe"]["units"]["target"]["aurabar"]["enable"] = false
 	E.db["unitframe"]["units"]["target"]["buffs"]["countFont"] = "Nimrod MT"
 	E.db["unitframe"]["units"]["target"]["buffs"]["yOffset"] = 15
@@ -367,11 +397,15 @@ function NUI:SetupLayout(layout)
 	E.db["unitframe"]["units"]["target"]["power"]["detachedWidth"] = 220
 	E.db["unitframe"]["units"]["target"]["power"]["enable"] = false
 	E.db["unitframe"]["units"]["target"]["width"] = 220
-	E.db["unitframe"]["units"]["targettarget"]["name"]["text_format"] = "[name:medium]"
+	E.db["unitframe"]["units"]["targettarget"]["height"] = 26
+	E.db["unitframe"]["units"]["targettarget"]["name"]["text_format"] = "[name:short]"
+	E.db["unitframe"]["units"]["targettarget"]["power"]["enable"] = false
 
 
 
-
+	--Niklas don't forget to handle movers
+	--if not E.db.movers then E.db.movers = {} end
+	--smiley
 
 	--[[
 		--If you want to modify the base layout according to