+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.1.1
+
+* Lots of changes to the ElvUI Layout.
+* Added Retail check for WindTools settings.
+
 ## Version 0.1
 
 * First private beta with Details, ElvUI, OmniCD, Plater, WarpDeplete profiles and some WeakAuras.